+++ NightgamesMod/nightgames/characters/custom/Stats.java
@@ -4,17 +4,16 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import nightgames.characters.Attribute;
 import nightgames.characters.Trait;
 
 public class Stats {
     public int level;
     public Map<Attribute, Integer> attributes;
-    public float stamina;
-    public float mojo;
-    public float arousal;
-    public float willpower;
+    public int stamina;
+    public int mojo;
+    public int arousal;
+    public int willpower;
     public List<Trait> traits;
 
     public Stats() {
